Fixed for test failure case when building for 64-bit. The define needs to be cast to the correct type.

+++ exampleProjects/IncrementalStore/Incremental Store Tests/PasswordTests.m
@@ -126,7 +126,7 @@ static NSString *const IncorrectPassword = @"IncorrectPassword";
     NSError *sqliteError = error.userInfo[NSUnderlyingErrorKey];
     STAssertNotNil(sqliteError, @"Nil SQLite error");
     STAssertEqualObjects(sqliteError.domain, NSSQLiteErrorDomain, @"Incorrect error SQLite error domain");
-    STAssertEquals(sqliteError.code, SQLITE_NOTADB, @"Incorrect error SQLite error code");
+    STAssertEquals(sqliteError.code, (NSInteger)SQLITE_NOTADB, @"Incorrect error SQLite error code");
     [self cleanUp:store];
     
     // Try again once more to be sure it still opens
